web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Javier Levio S"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.6"
dj-database-url = "^0.5.0"
python-decouple = "^3.3"
psycopg2-binary = "^2.8.5"
celery = "^4.4.5"
redis = "^3.5.3"
tqdm = "^4.46.1"
gunicorn = "^20.0.4"
djangorestframework = "^3.11.0"
flake8 = "^3.8.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"